==> 03_prep_DE_data.ipynb <==
# conda activate anndata

import numpy as np
import pandas as pd
import anndata as ad

Here I create pseduobulk data per donor and cell type to perform DE analysis for each cell type

import os
os.chdir("/mnt/lareaulab/reliscu/projects/NSF_GRFP/analyses/pseudobulk_test/tasic_2018/mouse_ALM")

adata = ad.read_h5ad("data/tasic_2018_ALM_STAR_model/tasic_2018_ALM_STAR_gene_counts_scVI.h5ad")

# # I want to find DE genes for each cell type. The best way to do this is by pseudobulking cell types by donor.

# # First let's see how the makeup of cell types break down by donor:

# pd.DataFrame(adata.obs.groupby("donor_id")["cell_subclass"].value_counts().groupby(level=0).head(1))

# Looks like donor is confounded with cell type (by design). 

# Let's try pseudobulking anyways.

ctype = "Pvalb"

adata_subset = adata[adata.obs['cell_subclass'] == ctype].copy()
X = adata_subset.raw.X

X.shape

df = pd.DataFrame.sparse.from_spmatrix(
    X, 
    index=adata_subset.obs_names,
    columns=adata.raw.var_names
)

df.head()

gene_cts = df.sum(axis=0)

df = df.loc[:, gene_cts > 0]

df.shape

corr_mat = np.corrcoef(df)
flat_idx = np.argmin(corr_mat)
row_idx, col_idx = np.unravel_index(flat_idx, corr_mat.shape)
print(row_idx, col_idx)

df.index[row_idx]

np.argmin(np.mean(corr_mat, axis=0))

np.mean(corr_mat, axis=0)[row_idx]

# Remove samples that are lowly correlated to other samples representing the same cell type

groups = df.groupby(adata_subset.obs['donor_id'], observed=True).groups

bulked = []
for donor, idx in groups.items():
    print(donor)
    df_subset = df.loc[idx]
    bulked.append(df_subset.sum())

column_names = list(groups.keys())
df_bulked = pd.concat(bulked, axis=1)

df_bulked.columns = column_names

df_bulked.head()

corr_mat = df_bulked.corr()   # shape: (n_cols, n_cols)
col_means = corr_mat.mean(axis=0)   # or axis=1, same here since it's symmetric
col_means.sort_values()

df_bulked_v2 = df.groupby(adata_subset.obs['donor_id']).sum()

df_bulked_v2.T.head()

corr_mat_pd = df_bulked_v2.T.corr()   # shape: (n_cols, n_cols)

col_means_pd = corr_mat_pd.mean(axis=1)   # or axis=1, same here since it's symmetric
col_means_pd.sort_values()

corr_mat_pd

df_list = []
meta_list = []

for ctype in np.unique(adata.obs['cell_subclass']):
    print(f"Starting {ctype}...")
    
    adata_subset = adata[adata.obs['cell_subclass'] == ctype].copy()
    
    X = adata_subset.raw.X
    
    df = pd.DataFrame.sparse.from_spmatrix(
        X, 
        index=adata_subset.obs_names,
        columns=adata.raw.var_names
    )
    df_bulked = df.groupby(adata_subset.obs['donor_id']).sum()
    
    meta_list.append(pd.DataFrame({
        'Cell_type': ctype, 
        'Donor': df_bulked.index.astype(str).values
    }))
    
    df_bulked.index = ctype + "_" + df_bulked.index.astype(str)
    df_list.append(df_bulked)

df_all = pd.concat(df_list, axis=0)
meta = pd.concat(meta_list, axis=0).reset_index(drop=True)

meta['Sample_ID'] = df_all.index.values
df_all.index.name = None

# Save
df_all.T.to_csv("data/tasic_2018_ALM_STAR_donor_cell_type_pseudobulk.csv")
meta.to_csv("data/tasic_2018_ALM_STAR_donor_cell_type_pseudobulk_sampleinfo.csv", index=False)
