==> 10_make_pseudobulk_PSI.ipynb <==
# conda activate anndata

import os
import sys

sys.path.append("/mnt/lareaulab/reliscu/code")

from junction2psi import *

pseudobulk_data = "SyntheticDataset1_25pcntCells_100SD_200samples_SJ_pseudobulk"
sj_file = f"data/tasic_2018_ALM_STAR_{pseudobulk_data}.csv"

min_observed = 0.05
# minPsi = 0.05

SJ_counts_table = pd.read_csv(sj_file, index_col=0)

events_i1 = pd.Index([x[:-3] for x in SJ_counts_table.index if '_I1' in x])
events_i2 = pd.Index([x[:-3] for x in SJ_counts_table.index if '_I2' in x])
events_se = pd.Index([x[:-3] for x in SJ_counts_table.index if '_SE' in x])

events = events_i1.intersection(events_i2).intersection(events_se)
i1_events = [x + '_I1' for x in events]
I1_table = SJ_counts_table.loc[i1_events]
I1_table.index = events

i2_events = [x + '_I2' for x in events]
I2_table = SJ_counts_table.loc[i2_events]
I2_table.index = events

se_events = [x + '_SE' for x in events]
SE_table = SJ_counts_table.loc[se_events]
SE_table.index = events

I1_filt = I1_table.index[I1_table.sum(axis=1) > 0]
I2_filt = I2_table.index[I2_table.sum(axis=1) > 0]
SE_filt = SE_table.index[SE_table.sum(axis=1) > 0]
filtered_events = I1_filt.intersection(I2_filt).intersection(SE_filt)

I1_table = I1_table.loc[filtered_events]
I2_table = I2_table.loc[filtered_events]
SE_table = SE_table.loc[filtered_events]

psi = ((I1_table + I2_table) /(2*SE_table + I1_table + I2_table)).fillna(0)
reads = SE_table + I1_table + I2_table

# alt_exons = psi.index[np.abs(0.5-psi.mean(axis=1)) <= (0.5-minPsi)]
obs_exons = psi.index[psi.isna().mean(axis=1) <= 1-min_observed]
# selected_exons = alt_exons.intersection(obs_exons)
selected_exons = obs_exons

len(selected_exons)

psi.shape[0]

psi = psi.loc[selected_exons]
reads = reads.loc[selected_exons]

psi.to_csv(f"data/tasic_2018_ALM_STAR_{pseudobulk_data}_min_observed{min_observed}_PSI.csv")
reads.to_csv(f"data/tasic_2018_ALM_STAR_{pseudobulk_data}_min_observed{min_observed}_counts.csv")
